Move get_stems_of
